pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dcc-mcp-photoshop"
version = "0.1.0"
description = "Adobe Photoshop adapter for the DCC Model Context Protocol (MCP) ecosystem — bridges MCP tools to Photoshop via UXP WebSocket"
authors = [{ name = "Long Hao", email = "hal.long@outlook.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
keywords = ["photoshop", "adobe", "mcp", "dcc", "ai", "llm", "model-context-protocol", "uxp"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "dcc-mcp-core>=0.12.18,<1.0.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.0",
    "pyfakefs>=5.3",
    "ruff>=0.8.0",
    "hatchling",
    "build",
    "pyyaml>=6.0",
    "pyinstaller>=6.0",
]

[project.urls]
Homepage = "https://github.com/loonghao/dcc-mcp-photoshop"
Repository = "https://github.com/loonghao/dcc-mcp-photoshop"
"Bug Tracker" = "https://github.com/loonghao/dcc-mcp-photoshop/issues"

[project.scripts]
dcc-mcp-photoshop = "dcc_mcp_photoshop.cli:main"

[project.entry-points."dcc_mcp.adapters"]
photoshop = "dcc_mcp_photoshop:PhotoshopMcpServer"

[tool.hatch.build.targets.wheel]
packages = ["src/dcc_mcp_photoshop"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP", "C4", "PLC"]
ignore = [
    "E501",   # line too long (handled by formatter)
    "UP007",  # Use X | Y for type annotations — requires Python 3.10+
    "UP006",  # Use type instead of Type — requires Python 3.9+
    "B008",   # Do not perform function calls in default arguments
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # assert is fine in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
